+++ examples/T-Block/Epaper_Badge/config.h
@@ -1,6 +1,6 @@
 // => Hardware select
-#define LILYGO_WATCH_BLOCK               // To use T-Watch Block , please uncomment this line
-// #define LILYGO_WATCH_BLOCK_V1             // To use T-Watch Block V1 , please uncomment this line
+// #define LILYGO_WATCH_BLOCK               // To use T-Watch Block , please uncomment this line
+#define LILYGO_WATCH_BLOCK_V1             // To use T-Watch Block V1 , please uncomment this line
 
 //NOT SUPPORT ...
 // #define LILYGO_WATCH_2020_V1
@@ -11,7 +11,7 @@
 // => Function select
 #define LILYGO_EINK_GDEH0154D67_BL             //It is electronic paper with backlight
 // #define LILYGO_EINK_GDEH0154D67_TP          //It is electronic paper with touchscreen
-// #define LILYGO_EINK_GDEP015OC1              //Single electronic paper, no backlight, no touch
+// #define LILYGO_EINK_GDEP015OC1              //Single electronic paper, no backlight, no touch -- Stop producing
 #include <LilyGoWatch.h>